add diagnostic service
